refactor(transaction): retire legacy event mapper name
