+++ tests/unit/services/calculators/portfolio_flow_bundle/test_portfolio_flow_bundle_slice0_characterization.py
@@ -19,7 +19,7 @@ from src.services.portfolio_transaction_processing_service.app.domain.position.r
 from src.services.portfolio_transaction_processing_service.app.infrastructure import (
     CashflowCalculator,
 )
-from src.services.portfolio_transaction_processing_service.app.infrastructure.legacy_transaction_event_mapper import (  # noqa: E501
+from src.services.portfolio_transaction_processing_service.app.infrastructure.booked_transaction_event_mapper import (  # noqa: E501
     to_booked_transaction,
 )
