+++ tests/unit/transaction_specs/test_interest_slice0_characterization.py
@@ -28,7 +28,7 @@ from src.services.portfolio_transaction_processing_service.app.infrastructure im
     CashflowCalculator,
     CostCalculationWorkflow,
 )
-from src.services.portfolio_transaction_processing_service.app.infrastructure.legacy_transaction_event_mapper import (  # noqa: E501
+from src.services.portfolio_transaction_processing_service.app.infrastructure.booked_transaction_event_mapper import (  # noqa: E501
     to_booked_transaction,
 )
 from src.services.query_service.app.dtos.transaction_dto import TransactionRecord
